+++ hypothesis-python/src/hypothesis/strategies/_internal/core.py
@@ -48,7 +48,6 @@ import attr
 
 from hypothesis._settings import note_deprecation
 from hypothesis.control import (
-    RandomSeeder,
     cleanup,
     current_build_context,
     deprecate_random_in_strategy,
@@ -983,8 +982,16 @@ def randoms(
     )
 
 
+class RandomSeeder:
+    def __init__(self, seed):
+        self.seed = seed
+
+    def __repr__(self):
+        return f"RandomSeeder({self.seed!r})"
+
+
 class RandomModule(SearchStrategy):
-    def do_draw(self, data):
+    def do_draw(self, data: ConjectureData) -> RandomSeeder:
         # It would be unsafe to do run this method more than once per test case,
         # because cleanup() runs tasks in FIFO order (at time of writing!).
         # Fortunately, the random_module() strategy wraps us in shared(), so
@@ -1022,6 +1029,7 @@ class BuildsStrategy(SearchStrategy[Ex]):
         args: tuple[SearchStrategy[Any], ...],
         kwargs: dict[str, SearchStrategy[Any]],
     ):
+        super().__init__()
         self.target = target
         self.args = args
         self.kwargs = kwargs
@@ -1063,7 +1071,7 @@ class BuildsStrategy(SearchStrategy[Ex]):
         current_build_context().record_call(obj, self.target, args, kwargs)
         return obj
 
-    def validate(self) -> None:
+    def do_validate(self) -> None:
         tuples(*self.args).validate()
         fixed_dictionaries(self.kwargs).validate()
 
@@ -1791,6 +1799,7 @@ def recursive(
 
 class PermutationStrategy(SearchStrategy):
     def __init__(self, values):
+        super().__init__()
         self.values = values
 
     def do_draw(self, data):
@@ -1821,6 +1830,7 @@ def permutations(values: Sequence[T]) -> SearchStrategy[list[T]]:
 
 class CompositeStrategy(SearchStrategy):
     def __init__(self, definition, args, kwargs):
+        super().__init__()
         self.definition = definition
         self.args = args
         self.kwargs = kwargs
@@ -2166,6 +2176,7 @@ def uuids(
 
 class RunnerStrategy(SearchStrategy):
     def __init__(self, default):
+        super().__init__()
         self.default = default
 
     def do_draw(self, data):