+++ hypothesis-python/tests/conjecture/test_data_tree.py
@@ -39,14 +39,14 @@ from tests.conjecture.common import (
     run_to_nodes,
 )
 
-TEST_SETTINGS = settings(
-    max_examples=5000, database=None, suppress_health_check=list(HealthCheck)
+runner_settings = settings(
+    max_examples=100, database=None, suppress_health_check=list(HealthCheck)
 )
 
 
 def runner_for(*examples):
     def accept(tf):
-        runner = ConjectureRunner(tf, settings=TEST_SETTINGS, random=Random(0))
+        runner = ConjectureRunner(tf, settings=runner_settings, random=Random(0))
         runner.exit_with = lambda reason: None
         ran_examples = []
         for choices in examples:
@@ -123,7 +123,9 @@ def test_novel_prefixes_are_novel():
             data.draw_bytes(1, 1, forced=b"\0")
             data.draw_integer(0, 3)
 
-    runner = ConjectureRunner(tf, settings=TEST_SETTINGS, random=Random(0))
+    runner = ConjectureRunner(
+        tf, settings=settings(runner_settings, max_examples=1000), random=Random(0)
+    )
     for _ in range(100):
         prefix = runner.tree.generate_novel_prefix(runner.random)
         extension = prefix + (ChoiceTemplate("simplest", count=100),)
@@ -135,7 +137,7 @@ def test_novel_prefixes_are_novel():
 def test_overruns_if_prefix():
     runner = ConjectureRunner(
         lambda data: [data.draw_boolean() for _ in range(2)],
-        settings=TEST_SETTINGS,
+        settings=runner_settings,
         random=Random(0),
     )
     runner.cached_test_function([False, False])
@@ -147,7 +149,7 @@ def test_stores_the_tree_flat_until_needed():
     def runner(data):
         for _ in range(10):
             data.draw_boolean()
-        data.mark_interesting()
+        data.mark_interesting(interesting_origin())
 
     root = runner.tree.root
     assert len(root.constraints) == 10
@@ -161,7 +163,7 @@ def test_split_in_the_middle():
         data.draw_integer(0, 1)
         data.draw_integer(0, 1)
         data.draw_integer(0, 15)
-        data.mark_interesting()
+        data.mark_interesting(interesting_origin())
 
     root = runner.tree.root
     assert len(root.constraints) == len(root.values) == 1
@@ -175,7 +177,7 @@ def test_stores_forced_nodes():
         data.draw_integer(0, 1, forced=0)
         data.draw_integer(0, 1)
         data.draw_integer(0, 1, forced=0)
-        data.mark_interesting()
+        data.mark_interesting(interesting_origin())
 
     root = runner.tree.root
     assert root.forced == {0, 2}
@@ -186,7 +188,7 @@ def test_correctly_relocates_forced_nodes():
     def runner(data):
         data.draw_integer(0, 1)
         data.draw_integer(0, 1, forced=0)
-        data.mark_interesting()
+        data.mark_interesting(interesting_origin())
 
     root = runner.tree.root
     assert root.transition.children[1].forced == {0}
@@ -496,7 +498,7 @@ def test_can_generate_hard_floats():
         def nodes(data):
             f = next_up_n(min_value, n)
             data.draw_float(min_value, max_value, forced=f, allow_nan=False)
-            data.mark_interesting()
+            data.mark_interesting(interesting_origin())
 
         data = ConjectureData.for_choices(
             [n.value for n in nodes], observer=tree.new_observer()